[refactor] AgilePriceCard: apply WidgetCard layout, simplified code, and rename to CurrentRateCard
